Don't have strings in plugins translatable that [dw]on't work anyhow
The plugin description and -path are not shown translated even if they
are (like for the helloqml plugin and German).
Showing as being translatable gives a wrong impression as examples for
other plugins.
BTW: Those descriptions do translate in the Plugin Editor/Creator's
console log, but only there.
While at it, make helloqml actually work, it apparently had never been
ported from 2.x to 3.x (and later).

Backport of #8985 plus some other random findings (not relevant to the
master branch)