Fix GH#10403: Moving notes between strings
Sort the notes by string, so that moved notes don't block each other.

Backport of #17410, convert `Selection::uniqueNotes()` from using `QList` to `std::list` in the due course, as that's how Mu4 does it too, and here avoids the use of the obsolete `QList::toStdList()`.
While at it, doing the same for `Selection::uniqueElements()`.