Allow Mu3 to open Mu4 mscz files
Assumption is that in a Mu4 mscz file the 1st mscx file is the real `rootfile`, the actual main score. In Mu3 and earlier it is the only `rootfile`, so this save there too.
In Mu4 pictures files are recorded as `rootfile`s too, so we read them in as such here, it there are any.
Except for `score_style.mss` and `viewsettings.json` (which both don't exist prior to Mu4 and are now read in if they are in the mscz), none of the other `rootfiles` are needed or usefull, as these represent the thumbnails (not read in Mu3 either), soundsettings (don't apply to Mu3) and the excerpts and their styles in Mu4.