Fix GH#261: Interfacing with musescore.com crashes program
by disabling WebEngine support.

Should also resolve the issue with a blank dialog when using Save online, like seen with the Windows 32-bit build as well as with the 3.6.2 Windows versions.

Force WebEngine off for MSVC too just like is was forced off for MinGW.
While for the latter is simply isn't available, for the former it doesn't work anymore since August '23, but causes a crash.

Disable WebEngine for Linux too, disabling 2 now failing vtests too. To fix those for real, they'd need to get regenerated by a linux build.