New build number format with month and day
Use month (01..12) and day (01..31) instead of day-of-year (001..366),
so we can more easily tell what date a given build was created on.

WiX requires a 32-bit integer, hence the new format requires us to lose
the final digit of the minute number (00..59), yielding a format like:

* `YYMMDDhhx` where `x` increments every 10 minutes (range 0..5)

This means we only know the build time to the nearest 10 minutes, but
that's OK since CI builds take longer than 10 minutes anyway.

Backport of #31564